+++ test/SoapClientFake.php
@@ -19,7 +19,7 @@ class SoapClientFake extends \SoapClient
      */
     public function __construct($wsdl, array $options = null)
     {
-        $wsdl = __DIR__ . '/Provider/_files/bcs-3.1.2/geschaeftskundenversand-api-3.1.2.wsdl';
+        $wsdl = __DIR__ . '/Provider/_files/bcs-3.3.2/geschaeftskundenversand-api-3.3.2.wsdl';
         parent::__construct($wsdl, $options);
     }
 }